added deprecate text comment, moved translations to a seperate file
