+++ packages/grafana-data/src/types/navModel.ts
@@ -1,6 +1,9 @@
 import { ComponentType } from 'react';
 
 export interface NavModelItem {
+  /** @deprecated NavModelItems are translated, so use the messages from navBarItem-translations.ts
+   *  instead and use item.text as a fallback only
+   */
   text: string;
   url?: string;
   subTitle?: string;
